+++ packages/rewild-renderer/types/interfaces.d.ts
@@ -1,0 +1,6 @@
+import { Renderer } from '../lib';
+
+interface IRenderable {
+  initialize(renderer: Renderer): Promise<IRenderable>;
+  render(renderer: Renderer, pass: GPURenderPassEncoder): void;
+}
